closes #56
