+++ .gitignore
@@ -98,3 +98,4 @@ local.properties
 **/build_errors.txt
 **/log.txt
 tmc-langs.log
+tmc-core.log*
