+++ src/main/java/fi/helsinki/cs/tmc/core/commands/Submit.java
@@ -36,7 +36,8 @@ public class Submit extends Command<SubmissionResult> {
      * {@code submissionPoller} for submitting the exercise located at {@code path}. The
      * {@code observer} will be notified of submission progress.
      */
-    public Submit(TmcSettings settings, ExerciseSubmitter submitter, SubmissionPoller submissionPoller, String path, ProgressObserver observer) {
+    public Submit(TmcSettings settings, ExerciseSubmitter submitter, 
+            SubmissionPoller submissionPoller, String path, ProgressObserver observer) {
         super(settings);
 
         this.path = path;
